// lib/api/generated/model/userResponseDtoStatus.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Uchat API
 * Uchat API Documentation
 * OpenAPI spec version: 1.0
 */

export type UserResponseDtoStatus = typeof UserResponseDtoStatus[keyof typeof UserResponseDtoStatus];


export const UserResponseDtoStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;